--- Server/routes/mailer.js
@@ -1,515 +1,0 @@
-var nodemailer = require('nodemailer');
-const express = require("express");
-const app = express();
-
-var transporter = nodemailer.createTransport({
-  service: 'gmail',
-  auth: {
-    user: 'mailmailer9000@gmail.com',
-    pass: '2021 pass'
-  }
-});
-
-
-
-app.get("/sendMail", async (req, res) => {
-    var mailOptions = {
-        from: 'mailmailer9000@gmail.com',
-        to: 'bullmeza1@gmail.com',
-        subject: 'Sending Email using Node.js',
-        html: 
-        `
-        <!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Transitional //EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-        <html xmlns="http://www.w3.org/1999/xhtml" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:o="urn:schemas-microsoft-com:office:office">
-        <head>
-        <!--[if gte mso 9]>
-        <xml>
-          <o:OfficeDocumentSettings>
-            <o:AllowPNG/>
-            <o:PixelsPerInch>96</o:PixelsPerInch>
-          </o:OfficeDocumentSettings>
-        </xml>
-        <![endif]-->
-          <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-          <meta name="viewport" content="width=device-width, initial-scale=1.0">
-          <meta name="x-apple-disable-message-reformatting">
-          <!--[if !mso]><!--><meta http-equiv="X-UA-Compatible" content="IE=edge"><!--<![endif]-->
-          <title></title>
-          
-            <style type="text/css">
-              table, td { color: #000000; } a { color: #0000ee; text-decoration: underline; }
-        @media only screen and (min-width: 620px) {
-          .u-row {
-            width: 600px !important;
-          }
-          .u-row .u-col {
-            vertical-align: top;
-          }
-        
-          .u-row .u-col-50 {
-            width: 300px !important;
-          }
-        
-          .u-row .u-col-100 {
-            width: 600px !important;
-          }
-        
-        }
-        
-        @media (max-width: 620px) {
-          .u-row-container {
-            max-width: 100% !important;
-            padding-left: 0px !important;
-            padding-right: 0px !important;
-          }
-          .u-row .u-col {
-            min-width: 320px !important;
-            max-width: 100% !important;
-            display: block !important;
-          }
-          .u-row {
-            width: calc(100% - 40px) !important;
-          }
-          .u-col {
-            width: 100% !important;
-          }
-          .u-col > div {
-            margin: 0 auto;
-          }
-        }
-        body {
-          margin: 0;
-          padding: 0;
-        }
-        
-        table,
-        tr,
-        td {
-          vertical-align: top;
-          border-collapse: collapse;
-        }
-        
-        p {
-          margin: 0;
-        }
-        
-        .ie-container table,
-        .mso-container table {
-          table-layout: fixed;
-        }
-        
-        * {
-          line-height: inherit;
-        }
-        
-        a[x-apple-data-detectors='true'] {
-          color: inherit !important;
-          text-decoration: none !important;
-        }
-        
-        </style>
-          
-          
-        
-        </head>
-        
-        <body class="clean-body" style="margin: 0;padding: 0;-webkit-text-size-adjust: 100%;background-color: #f9f9f9;color: #000000">
-          <!--[if IE]><div class="ie-container"><![endif]-->
-          <!--[if mso]><div class="mso-container"><![endif]-->
-          <table style="border-collapse: collapse;table-layout: fixed;border-spacing: 0;mso-table-lspace: 0pt;mso-table-rspace: 0pt;vertical-align: top;min-width: 320px;Margin: 0 auto;background-color: #f9f9f9;width:100%" cellpadding="0" cellspacing="0">
-          <tbody>
-          <tr style="vertical-align: top">
-            <td style="word-break: break-word;border-collapse: collapse !important;vertical-align: top">
-            <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td align="center" style="background-color: #f9f9f9;"><![endif]-->
-            
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: transparent;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: transparent;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="600" style="width: 600px;padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-100" style="max-width: 320px;min-width: 600px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-          <table height="0px" align="center" border="0" cellpadding="0" cellspacing="0" width="19%" style="border-collapse: collapse;table-layout: fixed;border-spacing: 0;mso-table-lspace: 0pt;mso-table-rspace: 0pt;vertical-align: top;border-top: 1px solid #f9f9f9;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%">
-            <tbody>
-              <tr style="vertical-align: top">
-                <td style="word-break: break-word;border-collapse: collapse !important;vertical-align: top;font-size: 0px;line-height: 0px;mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%">
-                  <span>&#160;</span>
-                </td>
-              </tr>
-            </tbody>
-          </table>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: transparent;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: transparent;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="600" style="width: 600px;padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-100" style="max-width: 320px;min-width: 600px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-        <table width="100%" cellpadding="0" cellspacing="0" border="0">
-          <tr>
-            <td style="padding-right: 0px;padding-left: 0px;" align="center">
-              
-              <img align="center" border="0" src="images/image-1.png" alt="Image" title="Image" style="outline: none;text-decoration: none;-ms-interpolation-mode: bicubic;clear: both;display: inline-block !important;border: none;height: auto;float: none;width: 29%;max-width: 168.2px;" width="168.2"/>
-              
-            </td>
-          </tr>
-        </table>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: transparent;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: transparent;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="600" style="width: 600px;padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-100" style="max-width: 320px;min-width: 600px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-          <table height="0px" align="center" border="0" cellpadding="0" cellspacing="0" width="100%" style="border-collapse: collapse;table-layout: fixed;border-spacing: 0;mso-table-lspace: 0pt;mso-table-rspace: 0pt;vertical-align: top;border-top: 1px solid #BBBBBB;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%">
-            <tbody>
-              <tr style="vertical-align: top">
-                <td style="word-break: break-word;border-collapse: collapse !important;vertical-align: top;font-size: 0px;line-height: 0px;mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%">
-                  <span>&#160;</span>
-                </td>
-              </tr>
-            </tbody>
-          </table>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: #ffffff;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: #ffffff;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="600" style="width: 600px;padding: 30px 0px 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-100" style="max-width: 320px;min-width: 600px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 30px 0px 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-          <div style="color: #3d3030; line-height: 140%; text-align: center; word-wrap: break-word;">
-            <p style="font-size: 14px; line-height: 140%; text-align: left;"><span style="font-size: 22px; line-height: 30.8px;">Hey, Robert wants to buy:</span></p>
-        <p style="font-size: 14px; line-height: 140%;">&nbsp;</p>
-          </div>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: transparent;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: transparent;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="600" style="width: 600px;padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-100" style="max-width: 320px;min-width: 600px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-        <table width="100%" cellpadding="0" cellspacing="0" border="0">
-          <tr>
-            <td style="padding-right: 0px;padding-left: 0px;" align="center">
-              
-              <img align="center" border="0" src="https://images-na.ssl-images-amazon.com/images/I/71BSd-0e5OL._AC_SX679_.jpg" alt="Image" title="Image" style="outline: none;text-decoration: none;-ms-interpolation-mode: bicubic;clear: both;display: inline-block !important;border: none;height: auto;float: none;width: 29%;max-width: 168.2px;" width="168.2"/>
-              
-            </td>
-          </tr>
-        </table>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: #ffffff;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: #ffffff;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="600" style="width: 600px;padding: 1px 0px 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-100" style="max-width: 320px;min-width: 600px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 1px 0px 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-          <div style="line-height: 140%; text-align: center; word-wrap: break-word;">
-            <p style="font-size: 14px; line-height: 140%;">Toy </p>
-          </div>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: transparent;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: transparent;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="600" style="width: 600px;padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-100" style="max-width: 320px;min-width: 600px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:0px 10px;font-family:georgia,palatino;" align="left">
-                
-          <div style="color: #000000; line-height: 130%; text-align: center; word-wrap: break-word;">
-            <p style="font-size: 14px; line-height: 130%;"><span style="font-size: 14px; line-height: 18.2px;"><span style="line-height: 18.2px; font-size: 14px;">$99</span></span></p>
-          </div>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-          <table height="0px" align="center" border="0" cellpadding="0" cellspacing="0" width="100%" style="border-collapse: collapse;table-layout: fixed;border-spacing: 0;mso-table-lspace: 0pt;mso-table-rspace: 0pt;vertical-align: top;border-top: 1px solid #BBBBBB;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%">
-            <tbody>
-              <tr style="vertical-align: top">
-                <td style="word-break: break-word;border-collapse: collapse !important;vertical-align: top;font-size: 0px;line-height: 0px;mso-line-height-rule: exactly;-ms-text-size-adjust: 100%;-webkit-text-size-adjust: 100%">
-                  <span>&#160;</span>
-                </td>
-              </tr>
-            </tbody>
-          </table>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-        
-        <div class="u-row-container" style="padding: 0px;background-color: transparent">
-          <div class="u-row" style="Margin: 0 auto;min-width: 320px;max-width: 600px;overflow-wrap: break-word;word-wrap: break-word;word-break: break-word;background-color: transparent;">
-            <div style="border-collapse: collapse;display: table;width: 100%;background-color: transparent;">
-              <!--[if (mso)|(IE)]><table width="100%" cellpadding="0" cellspacing="0" border="0"><tr><td style="padding: 0px;background-color: transparent;" align="center"><table cellpadding="0" cellspacing="0" border="0" style="width:600px;"><tr style="background-color: transparent;"><![endif]-->
-              
-        <!--[if (mso)|(IE)]><td align="center" width="300" style="width: 300px;padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-50" style="max-width: 320px;min-width: 300px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-        <div align="center">
-          <!--[if mso]><table width="100%" cellpadding="0" cellspacing="0" border="0" style="border-spacing: 0; border-collapse: collapse; mso-table-lspace:0pt; mso-table-rspace:0pt;font-family:georgia,palatino;"><tr><td style="font-family:georgia,palatino;" align="center"><v:roundrect xmlns:v="urn:schemas-microsoft-com:vml" xmlns:w="urn:schemas-microsoft-com:office:word" href="" style="height:36px; v-text-anchor:middle; width:280px;" arcsize="11%" stroke="f" fillcolor="#2dc26b"><w:anchorlock/><center style="color:#FFFFFF;font-family:georgia,palatino;"><![endif]-->
-            <a href="" target="_blank" style="box-sizing: border-box;display: inline-block;font-family:georgia,palatino;text-decoration: none;-webkit-text-size-adjust: none;text-align: center;color: #FFFFFF; background-color: #2dc26b; border-radius: 4px; -webkit-border-radius: 4px; -moz-border-radius: 4px; width:100%; max-width:100%; overflow-wrap: break-word; word-break: break-word; word-wrap:break-word; mso-border-alt: none;">
-              <span style="display:block;padding:10px 20px;line-height:120%;"><span style="font-size: 14px; line-height: 16.8px;">Confirm Purchase</span></span>
-            </a>
-          <!--[if mso]></center></v:roundrect></td></tr></table><![endif]-->
-        </div>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-        <!--[if (mso)|(IE)]><td align="center" width="300" style="width: 300px;padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;" valign="top"><![endif]-->
-        <div class="u-col u-col-50" style="max-width: 320px;min-width: 300px;display: table-cell;vertical-align: top;">
-          <div style="width: 100% !important;">
-          <!--[if (!mso)&(!IE)]><!--><div style="padding: 0px;border-top: 0px solid transparent;border-left: 0px solid transparent;border-right: 0px solid transparent;border-bottom: 0px solid transparent;"><!--<![endif]-->
-          
-        <table style="font-family:georgia,palatino;" role="presentation" cellpadding="0" cellspacing="0" width="100%" border="0">
-          <tbody>
-            <tr>
-              <td style="overflow-wrap:break-word;word-break:break-word;padding:10px;font-family:georgia,palatino;" align="left">
-                
-        <div align="center">
-          <!--[if mso]><table width="100%" cellpadding="0" cellspacing="0" border="0" style="border-spacing: 0; border-collapse: collapse; mso-table-lspace:0pt; mso-table-rspace:0pt;font-family:georgia,palatino;"><tr><td style="font-family:georgia,palatino;" align="center"><v:roundrect xmlns:v="urn:schemas-microsoft-com:vml" xmlns:w="urn:schemas-microsoft-com:office:word" href="" style="height:36px; v-text-anchor:middle; width:280px;" arcsize="11%" stroke="f" fillcolor="#e03e2d"><w:anchorlock/><center style="color:#FFFFFF;font-family:georgia,palatino;"><![endif]-->
-            <a href="" target="_blank" style="box-sizing: border-box;display: inline-block;font-family:georgia,palatino;text-decoration: none;-webkit-text-size-adjust: none;text-align: center;color: #FFFFFF; background-color: #e03e2d; border-radius: 4px; -webkit-border-radius: 4px; -moz-border-radius: 4px; width:100%; max-width:100%; overflow-wrap: break-word; word-break: break-word; word-wrap:break-word; mso-border-alt: none;">
-              <span style="display:block;padding:10px 20px;line-height:120%;"><span style="font-size: 14px; line-height: 16.8px;">Cancel Purchase</span></span>
-            </a>
-          <!--[if mso]></center></v:roundrect></td></tr></table><![endif]-->
-        </div>
-        
-              </td>
-            </tr>
-          </tbody>
-        </table>
-        
-          <!--[if (!mso)&(!IE)]><!--></div><!--<![endif]-->
-          </div>
-        </div>
-        <!--[if (mso)|(IE)]></td><![endif]-->
-              <!--[if (mso)|(IE)]></tr></table></td></tr></table><![endif]-->
-            </div>
-          </div>
-        </div>
-        
-        
-            <!--[if (mso)|(IE)]></td></tr></table><![endif]-->
-            </td>
-          </tr>
-          </tbody>
-          </table>
-          <!--[if mso]></div><![endif]-->
-          <!--[if IE]></div><![endif]-->
-        </body>
-        
-        </html>
-        
-        `
-      };
-      
-
-    transporter.sendMail(mailOptions, function(error, info){
-        if (error) {
-            console.log(error);
-        } else {
-            res.status(200).send(info.response);
-        }
-      });
-});
-
-
-app.get("/amazonRedirect", (req, res) => {
-    /*
-    Send user to "new page"
-    new page gets their session id, subtracts the money, and redirects to amazon
-    */
-});
-
-
-module.exports = app;
-
